chore(api): Phase 2 - Auth module migration (pilot)
- Create utils/api/auth.ts with all auth functions:
  * loginUser
  * verifyEmail
  * resendVerificationEmail
  * registerChild
  * registerUser
- Update utils/api/index.ts to export auth functions
- Update utils/api.ts to re-export from new auth module
- Remove original auth function implementations (322 lines)

Pilot migration successful - no breaking changes
